+++ src/__internal__/input/input-style-overrides.style.ts
@@ -321,7 +321,7 @@ export const selectStyleOverrides = (
 export const numeralDateStyles = css`
   .numeral-date-wrapper & {
     .input-text-container input {
-      text-align: center;
+      text-align: left;
     }
   }
 `;