+++ src/__internal__/popover-menu/menu-item/menu-item-content/menu-item-leading.component.tsx
@@ -1,5 +1,6 @@
-import React from "react";
+import React, { useContext } from "react";
 import styled from "styled-components";
+import { PopoverMenuContext } from "../../contexts";
 import Icon from "../../../../components/icon";
 
 const StyledMenuItemLeading = styled.span`
@@ -14,6 +15,7 @@ const StyledMenuItemLeading = styled.span`
 
 const StyledSelectedIconWrapper = styled.span<{ $hasIcon: boolean }>`
   ${({ $hasIcon }) => `
+    padding: 0 var(--global-space-comp-2-xs);
     visibility: ${$hasIcon ? "visible" : "hidden"};
   `}
 `;
@@ -25,14 +27,17 @@ const MenuItemLeading = ({
   children?: React.ReactNode;
   selectedIcon?: boolean;
 }) => {
+  const { isButtonMenu } = useContext(PopoverMenuContext);
   return (
     <StyledMenuItemLeading className="menu-item-leading" data-element="leading">
-      <StyledSelectedIconWrapper
-        data-role="selected-icon-wrapper"
-        $hasIcon={selectedIcon}
-      >
-        <Icon type="tick_thick" data-role="selected-icon" />
-      </StyledSelectedIconWrapper>
+      {!isButtonMenu && (
+        <StyledSelectedIconWrapper
+          data-role="selected-icon-wrapper"
+          $hasIcon={selectedIcon}
+        >
+          <Icon type="tick_thick" data-role="selected-icon" />
+        </StyledSelectedIconWrapper>
+      )}
       {children}
     </StyledMenuItemLeading>
   );